tmp improvements
