+++ src/simtools/corsika/corsika_output_validator.py
@@ -24,8 +24,8 @@ def validate_corsika_output(data_files, log_files, expected_shower_events=None, 
     curved_atmo: bool, optional
         Whether the CORSIKA simulation was run with the curved atmosphere option.
     """
-    data_files = general.ensure_iterable(data_files) if data_files is not None else []
-    log_files = general.ensure_iterable(log_files)
+    data_files = general.ensure_list(data_files) if data_files is not None else []
+    log_files = general.ensure_list(log_files)
 
     validate_event_numbers(data_files, expected_shower_events)
     validate_log_files(
